master/requirements.txt
python-socketio==3.1.2
aiohttp==3.5.4